A base use cases document for the DIEM WG
This is very lightly edited, and has authorship changes, from the expired, pre-charter, document draft-haberman-digital-emblem-ps-03.  The edits are in the abstract to start an alignment with the charter, but a lot of alignment work is needed - see, and make, PRs, for this.  A goal is to preserve the work on extensive use cases, while restructuring the document so that the main body will cover only use cases in the initial scope of the WG and the requirements of those use cases.